requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
distlib==0.3.4
Django~=4.0.1
django-cors-headers==3.10.1
djangorestframework==3.13.1
filelock==3.4.2
Pillow==8.4.0
platformdirs==2.4.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.10.0
whitenoise==5.3.0
jwt==1.3.1
PyJWT==2.3.0